# CAMEL-19295: Basic thread-safe LRU cache