… core/camel-core/src/test/java/org/apache/camel/support/DefaultLRUCacheFactoryTest.java