@davsclaus 2023-05-23
Yeah can you test with camel-caffeine-lrucache as well.

Also I think we can have a non LRU cache, we often just want to keep 1000 elements in a cache, and they may not need to be exact LRU based. So they can be FIFO or anything like that, just that the size of big enough for normal use-cases.

So basically we can make a cache that is just a ConcurrentMap from the JDK to be used in the various places, where a LRU is really not needed.